requirements.txt
streamlit>=1.46.0
python-dotenv>=1.1.0
sentence-transformers>=2.7.0
faiss-cpu>=1.11.0
langchain>=0.3.26
langchain-community>=0.3.26
google-genai>=1.21.1
PyPDF2>=3.0.1
python-docx>=1.2.0
numpy>=2.3.1